+++ src/LiveComponent/tests/Fixtures/Component/TrackRenders.php
@@ -1,0 +1,32 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\UX\LiveComponent\Tests\Fixtures\Component;
+
+use Symfony\UX\LiveComponent\Attribute\AsLiveComponent;
+use Symfony\UX\LiveComponent\Attribute\LiveProp;
+use Symfony\UX\LiveComponent\Attribute\PreReRender;
+use Symfony\UX\LiveComponent\DefaultActionTrait;
+
+#[AsLiveComponent('track_renders')]
+final class TrackRenders
+{
+    use DefaultActionTrait;
+
+    #[LiveProp]
+    public int $reRenders = 0;
+
+    #[PreReRender]
+    public function preReRender(): void
+    {
+        ++$this->reRenders;
+    }
+}