+++ src/TwigComponent/src/Test/RenderedComponent.php
@@ -11,6 +11,8 @@
 
 namespace Symfony\UX\TwigComponent\Test;
 
+use Symfony\Component\DomCrawler\Crawler;
+
 /**
  * @author Kevin Bond <kevinbond@gmail.com>
  */
@@ -21,6 +23,20 @@ final class RenderedComponent implements \Stringable
      */
     public function __construct(private string $html)
     {
+    }
+
+    public function crawler(): Crawler
+    {
+        if (!class_exists(Crawler::class)) {
+            throw new \LogicException(sprintf('"symfony/dom-crawler" is required to use "%s()" (install with "composer require symfony/dom-crawler").', __METHOD__));
+        }
+
+        return new Crawler($this->html);
+    }
+
+    public function toString(): string
+    {
+        return $this->html;
     }
 
     public function __toString(): string